More graphql work

+++ src/graphql/type/account.type.ts
@@ -22,7 +22,7 @@ export class AccountType implements IGraphQLType {
         accessLevel: Int
         organization: String
         permissions: [String]
-        registeredOn: Int
+        registeredOn: Float
         registeredBy: String
       }
